move all tests related to formatting evidence/alerts text to test_evidence_formatter.py
